@@ -4,10 +4,13 @@
    "description": "Progress Pioneer package.",
    "devDependencies": {
       "@types/express": "^4.17.20",
+      "@types/jest": "^29.5.8",
       "@types/node": "^16.18.61",
+      "jest": "^29.7.0",
       "nodemon": "^3.0.1",
+      "ts-jest": "^29.1.1",
       "ts-node": "^10.7.0",
-      "typescript": "^4.5.2"
+      "typescript": "^4.9.5"
    },
    "dependencies": {
       "express": "^4.18.2",
@@ -17,6 +20,8 @@
    },
    "scripts": {
       "start": "nodemon",
+      "test": "jest --watch",
+      "build": "tsc",
       "migration": "typeorm migration:run -d ./src/data-source.js"
    }
 }
